edited mfhd and added traf box
Edited mfhd to make sure that SetSize converts s it uint32 when s less
then uint32. Added struct and methods to traf box.

+++ mpeg4file/moof/mfhd.go
@@ -29,7 +29,7 @@ func (m* mfhd) SetSize(s uint64){
 		m.size = 1
 		m.largeSize = s
 	} else {
-		m.size = s
+		m.size = uint32(s)
 	}
 }
